backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==2.10.0
python-jose==3.3.0
cryptography==41.0.7
requests==2.32.3
pytest==7.4.0
httpx==0.27.0

# Phase 3 ML Dependencies
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3

# Phase 5: Rate limiting
slowapi==0.1.9
